Split out Mypy-dependant typing examples, add more type checkers
