+++ tox.ini
@@ -1,3 +1,5 @@
+# SPDX-License-Identifier: MIT
+
 [tox]
 min_version = 4
 env_list =
@@ -5,7 +7,8 @@ env_list =
     py3{9,10,11,12,13,14}-tests,
     py3{10,11,12,13,14}-mypy,
     pypy3-tests,
-    pyright,
+    # Mypy needs to run within the respective Python version
+    typing-{pyright,ty,pyrefly}
     docs-{sponsors,doctests},
     changelog,
     coverage-combine,
@@ -26,7 +29,7 @@ dependency_groups =
 commands =
     tests: pytest {posargs:-n auto}
     mypy: pytest -k test_mypy
-    mypy: mypy tests/typing_example.py
+    mypy: mypy typing-examples
     mypy: mypy src/attrs/__init__.pyi src/attr/__init__.pyi src/attr/_typing_compat.pyi src/attr/_version_info.pyi src/attr/converters.pyi src/attr/exceptions.pyi src/attr/filters.pyi src/attr/setters.pyi src/attr/validators.pyi
 
 [testenv:pypy3-tests]
@@ -121,9 +124,19 @@ commands =
     towncrier build --version main --draft
 
 
-[testenv:pyright]
+[testenv:typing-pyright]
 dependency_groups = pyright
-commands = pytest tests/test_pyright.py -vv
+commands =
+    pyright typing-examples/baseline_examples.py
+    pytest tests/test_pyright.py -vv
+
+[testenv:typing-ty]
+dependency_groups = ty
+commands = ty check typing-examples/baseline_examples.py
+
+[testenv:typing-pyrefly]
+dependency_groups = pyrefly
+commands = pyrefly check typing-examples/baseline_examples.py
 
 
 [testenv:docset]
